[experimental] Reorganize JSON-RPC code into subfolder ## Summary
Doing this, so that it can have its own `__tests__` folder, into which we'll put all the RPC method live tests.
